+++ src/main/java/com/project/plateforme_dannotation_collaborative/Dto/Admin/DatasetMinResposeDto.java
@@ -14,7 +14,7 @@ public class DatasetMinResposeDto  {
     Long id;
     String name;
     String description;
-    Double advancement ;
+    Integer advancement ;
     Boolean annotated ;
     Integer size ;
     Integer numberClasses;
